refactor(api): split routes/mod.rs into per-resource submodules (#477)
routes/mod.rs held 2931 lines of handler code. Move it into eleven named
siblings (auth, users, config_io, status, vpn, interfaces, settings, ai,
rbac, response) leaving mod.rs as declarations, re-exports, and the shared
status-code helpers. Pure relocation — the `pub use <mod>::*` re-exports
keep every existing `routes::<handler>` path in main.rs resolving, so no
call sites change.

Line-for-line identical code (verified); `extract_user_id` gains
`pub(super)` so routes::users can share it. cargo check clean, 639 tests
green.

Bump 5.97.12 -> 5.97.13.
